Вынесены из основного контроллера структуры: - QuizStepViewModel - QuizResultViewModel - QuizQuestion
Создана QuestionFabric и в неё перенесён массив вопросов и создан метод выбирающий рандомный вопрос из массива.
