+++ MovieQuiz/Models/QuizQuestion.swift
@@ -1,0 +1,11 @@
+//
+// Model данных для вопросов
+//
+
+import Foundation
+
+private struct QuizQuestion {
+    let image: String
+    let text: String
+    let correctAnswer: Bool
+}
