+++ MovieQuiz/Models/QuizStepViewModel.swift
@@ -1,0 +1,11 @@
+//
+// ViewModel отдельного вопроса
+//
+
+import UIKit
+
+private struct QuizStepViewModel {
+    let image: UIImage
+    let question: String
+    let questionNumber: String
+}
